+++ pinger_mock.go
@@ -1,0 +1,9 @@
+package healthy
+
+type MockPinger struct {
+	Err error
+}
+
+func (mp *MockPinger) Ping() error {
+	return mp.Err
+}
